[FEAT] Move Transaction Event from BottomSheetShareFragment to NoCategoryFragment (#149)
