+++ app/src/main/java/org/sopt/havit/ui/share/BottomSheetShareFragment.kt
@@ -7,17 +7,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.lifecycleScope
-import androidx.navigation.fragment.NavHostFragment
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import kotlinx.coroutines.*
+import kotlinx.coroutines.launch
 import org.sopt.havit.R
 import org.sopt.havit.data.RetrofitObject
 import org.sopt.havit.databinding.FragmentBottomSheetShareBinding
 import org.sopt.havit.util.MySharedPreference
-import retrofit2.Call
-import java.lang.Exception
 
 class BottomSheetShareFragment : BottomSheetDialogFragment() {
     private var _binding: FragmentBottomSheetShareBinding? = null
@@ -27,13 +24,12 @@ class BottomSheetShareFragment : BottomSheetDialogFragment() {
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         return BottomSheetDialog(requireContext(), R.style.AppBottomSheetDialogTheme)
     }
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         _binding = FragmentBottomSheetShareBinding.inflate(layoutInflater, container, false)
-
-        getCategoryNum()
 
         return binding.root
     }
@@ -43,33 +39,6 @@ class BottomSheetShareFragment : BottomSheetDialogFragment() {
         setBottomSheetHeight()
     }
 
-
-    private fun getCategoryNum() {
-        var num = 0
-        lifecycleScope.launch {
-            try {
-                val response =
-                    RetrofitObject.provideHavitApi(MySharedPreference.getXAuthToken(requireContext()))
-                        .getCategoryNum()
-                num = response.data.size
-                initTransaction(num)
-                Log.d("BottomSheetShareFragment", num.toString())
-            } catch (e: Exception) {
-            }
-        }
-    }
-
-    private fun initTransaction(categoryNum : Int) {
-
-        Log.d("BottomSheetShareFragment_Transaction" , categoryNum.toString())
-
-        // 이미 생성된 카테고리가 있으면 카테고리 선택 Fragment 로 이동
-        if ( categoryNum!= 0) {
-            childFragmentManager.beginTransaction()
-                .replace(R.id.fcv_share, SelectCategoryFragment())
-                .commit()
-        }
-    }
 
     private fun setBottomSheetHeight() {
         (dialog as BottomSheetDialog).behavior.apply {
