+++ app/src/main/java/org/sopt/havit/ui/share/NoCategoryFragment.kt
@@ -1,13 +1,18 @@
 package org.sopt.havit.ui.share
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
+import kotlinx.coroutines.launch
 import org.sopt.havit.R
+import org.sopt.havit.data.RetrofitObject
 import org.sopt.havit.databinding.FragmentNoCategoryBinding
+import org.sopt.havit.util.MySharedPreference
 
 class NoCategoryFragment : Fragment() {
 
@@ -20,12 +25,32 @@ class NoCategoryFragment : Fragment() {
     ): View {
         _binding = FragmentNoCategoryBinding.inflate(layoutInflater, container, false)
 
+        getCategoryNum()
         initClickListener()
 
         return binding.root
     }
 
-    private fun initClickListener(){
+    private fun getCategoryNum() {
+        lifecycleScope.launch {
+            try {
+                val response =
+                    RetrofitObject.provideHavitApi(MySharedPreference.getXAuthToken(requireContext()))
+                        .getCategoryNum()
+                val num = response.data.size
+                initTransaction(num)
+                Log.d("BottomSheetShareFragment", num.toString())
+            } catch (e: Exception) {
+            }
+        }
+    }
+
+    private fun initTransaction(categoryNum: Int) {
+        if (categoryNum != 0)
+            findNavController().navigate(R.id.action_noCategoryFragment_to_selectCategoryFragment)
+    }
+
+    private fun initClickListener() {
         binding.btnAddCategory.setOnClickListener {
             findNavController().navigate(R.id.action_noCategoryFragment_to_addCategoryFragment)
         }
